moved things to VariantContextGetters class
