+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/validation/CalculateMixingFractions.java
@@ -7,10 +7,10 @@ import org.broadinstitute.barclay.argparser.Argument;
 import org.broadinstitute.barclay.argparser.CommandLineProgramProperties;
 import org.broadinstitute.barclay.help.DocumentedFeature;
 import org.broadinstitute.hellbender.cmdline.StandardArgumentDefinitions;
+import org.broadinstitute.hellbender.utils.variant.VariantContextGetters;
 import picard.cmdline.programgroups.VariantEvaluationProgramGroup;
 import org.broadinstitute.hellbender.engine.*;
 
-import org.broadinstitute.hellbender.utils.GATKProtectedVariantContextUtils;
 import org.broadinstitute.hellbender.utils.locusiterator.AlignmentStateMachine;
 import org.broadinstitute.hellbender.utils.pileup.ReadPileup;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
@@ -148,6 +148,6 @@ public class CalculateMixingFractions extends VariantWalker {
     }
 
     private static double[] getArrayAttribute(final VariantContext vc, final String attribute) {
-        return GATKProtectedVariantContextUtils.getAttributeAsDoubleArray(vc, attribute, () -> null, -1);
+        return VariantContextGetters.getAttributeAsDoubleArray(vc, attribute, () -> null, -1);
     }
 }
