+++ src/main/java/org/broadinstitute/hellbender/utils/GATKProtectedVariantContextUtils.java
@@ -3,27 +3,21 @@ package org.broadinstitute.hellbender.utils;
 import htsjdk.samtools.util.Locatable;
 import htsjdk.variant.variantcontext.Allele;
 import htsjdk.variant.variantcontext.Genotype;
-import htsjdk.variant.variantcontext.GenotypeBuilder;
 import htsjdk.variant.variantcontext.VariantContext;
 import htsjdk.variant.vcf.VCFConstants;
 import org.apache.commons.lang3.ArrayUtils;
 import org.broadinstitute.hellbender.engine.filters.ReadFilterLibrary;
-import org.broadinstitute.hellbender.exceptions.GATKException;
 import org.broadinstitute.hellbender.utils.locusiterator.AlignmentStateMachine;
 import org.broadinstitute.hellbender.utils.param.ParamUtils;
 import org.broadinstitute.hellbender.utils.pileup.PileupElement;
 import org.broadinstitute.hellbender.utils.pileup.ReadPileup;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
+import org.broadinstitute.hellbender.utils.variant.VariantContextGetters;
 
-import java.lang.reflect.Array;
 import java.util.*;
-import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.function.ToDoubleFunction;
-import java.util.function.ToIntFunction;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 /**
@@ -31,341 +25,6 @@ import java.util.stream.StreamSupport;
  */
 public class GATKProtectedVariantContextUtils {
 
-    /**
-     * Composes the double array from a genotype annotation.
-     *
-     * @param variantContext the target variant-context.
-     * @param key the name of the attribute containing the double array.
-     * @param defaultValue the double array to return in case there is no such an annotation.
-     * @param missingValue value to use to fill up positions with a missing value (e.g. '.').
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static double[] getAttributeAsDoubleArray(final VariantContext variantContext, final String key,
-                                                     final Supplier<double[]> defaultValue, final double missingValue) {
-        Utils.nonNull(variantContext);
-        return attributeValueToDoubleArray(variantContext.getAttribute(key), key, defaultValue, missingValue);
-    }
-
-    /**
-     * Composes the int array from an INFO annotation.
-     *
-     * @param variantContext the target variant-context.
-     * @param key the name of the attribute containing the double array.
-     * @param defaultValue the double array to return in case there is no such an annotation.
-     * @param missingValue value to use to fill up positions with a missing value (e.g. '.').
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static int[] getAttributeAsIntArray(final VariantContext variantContext, final String key,
-                                                     final Supplier<int[]> defaultValue, final int missingValue) {
-        Utils.nonNull(variantContext);
-        return attributeValueToIntArray(variantContext.getAttribute(key), key, defaultValue, missingValue);
-    }
-
-    /**
-     * Composes the double array from a genotype annotation. Provides default and missing values.
-     *
-     * @param variantContext the target variant-context.
-     * @param attribute the name of the attribute containing the double array.
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static double[] getAttributeAsDoubleArray(final VariantContext variantContext, final String attribute) {
-        return getAttributeAsDoubleArray(variantContext, attribute, () -> null, -1);
-    }
-
-    /**
-     * Composes the double array from a genotype annotation.
-     *
-     * @param genotype the target variant-context.
-     * @param key the name of the attribute containing the double array.
-     * @param defaultValue the double array to return in case there is no such an annotation.
-     * @param missingValue value to use to fill up positions with a missing value (e.g. '.').
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static double[] getAttributeAsDoubleArray(final Genotype genotype, final String key,
-                                                     final Supplier<double[]> defaultValue, final double missingValue) {
-        Utils.nonNull(genotype);
-        return attributeValueToDoubleArray(genotype.getExtendedAttribute(key), key, defaultValue, missingValue);
-    }
-
-    /**
-     * Composes the double array from a genotype annotation.
-     *
-     * @param genotype the target variant-context.
-     * @param key the name of the attribute containing the double array.
-     * @param defaultValue the double array to return in case there is no such an annotation.
-     * @param missingValue value to use to fill up positions with a missing value (e.g. '.').
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static int[] getAttributeAsIntArray(final Genotype genotype, final String key,
-                                                     final Supplier<int[]> defaultValue, final int missingValue) {
-        Utils.nonNull(genotype);
-        return attributeValueToIntArray(genotype.getExtendedAttribute(key), key, defaultValue, missingValue);
-    }
-
-
-    /**
-     * Get Long attribute from a variant context.
-     *
-     * @param variantContext the target variant-context.
-     * @param attribute the name of the attribute containing the Long value.
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static Long getAttributeAsLong(final VariantContext variantContext, final String attribute, final Long defaultValue) {
-        Utils.nonNull(variantContext);
-        Utils.nonNull(attribute);
-        Object x = variantContext.getAttribute(attribute);
-        if ( x == null || x.equals(VCFConstants.MISSING_VALUE_v4) ) return defaultValue;
-        if ( x instanceof Number ) return ((Number) x).longValue();
-        return Long.valueOf((String)x); // throws an exception if this isn't a string
-    }
-
-    /**
-     * Composes the Long List from a variant context.
-     *
-     * @param variantContext the target variant-context.
-     * @param attribute the name of the attribute containing the Long list.
-     * @return never {@code null}.
-     * @throws IllegalArgumentException if {@code variantContext} is {@code null} or {@code key} is {@code null}.
-     */
-    public static List<Long> getAttributeAsLongList(final VariantContext variantContext, final String attribute, final Long defaultValue) {
-        Utils.nonNull(variantContext);
-        Utils.nonNull(attribute);
-        return variantContext.getAttributeAsList(attribute).stream().map(
-                x -> {
-                    if (x == null || x.equals(VCFConstants.MISSING_VALUE_v4)) {
-                        return defaultValue;
-                    } else if (x instanceof Number) {
-                        return ((Number) x).longValue();
-                    } else {
-                        return Long.valueOf((String)x); // throws an exception if this isn't a string
-                    }
-                }
-        ).collect(Collectors.toList());
-    }
-
-    private static double[] attributeValueToDoubleArray(final Object value, final String key, final Supplier<double[]> defaultResult, final double missingValue) {
-        Utils.nonNull(key);
-        final ToDoubleFunction<Object> doubleConverter = o -> {
-            if (o == null) {
-                return missingValue;
-            } else {
-                final String s = String.valueOf(o);
-                if (s.equals(VCFConstants.MISSING_VALUE_v4)) {
-                    return missingValue;
-                } else {
-                    try {
-                        return Double.parseDouble(s);
-                    } catch (final NumberFormatException ex) {
-                        throw new GATKException(String.format("INFO annotation '%s' contains a non-double value '%s'", key, s), ex);
-                    }
-                }
-            }
-        };
-
-        if (value == null) {
-            return defaultResult.get();
-        } else if (value.getClass().isArray()) {
-            final double[] result = new double[Array.getLength(value)];
-            for (int i = 0; i < result.length; i++) {
-                result[i] = doubleConverter.applyAsDouble(String.valueOf(Array.get(value, i)));
-            }
-            return result;
-        } else if (value.getClass().isAssignableFrom(Iterable.class)) {
-            return StreamSupport.stream(((Iterable<?>)value).spliterator(), false)
-                    .mapToDouble(doubleConverter).toArray();
-        } else { // as a last resort with transform it into an String and try to parse an array out of it.
-            return Stream.of(String.valueOf(value).trim().replaceAll("\\[|\\]", "")
-                    .split(VCFConstants.INFO_FIELD_ARRAY_SEPARATOR))
-                    .mapToDouble(doubleConverter).toArray();
-        }
-    }
-
-    private static int[] attributeValueToIntArray(final Object value, final String key, final Supplier<int[]> defaultResult, final int missingValue) {
-        Utils.nonNull(key);
-        final ToIntFunction<Object> intConverter = o -> {
-            if (o == null) {
-                return missingValue;
-            } else {
-                final String s = String.valueOf(o).trim();
-                if (s.equals(VCFConstants.MISSING_VALUE_v4)) {
-                    return missingValue;
-                } else {
-                    try {
-                        return Integer.parseInt(s);
-                    } catch (final NumberFormatException ex) {
-                        throw new GATKException(String.format("INFO annotation '%s' contains a non-int value '%s'", key, s), ex);
-                    }
-                }
-            }
-        };
-
-        if (value == null) {
-            return defaultResult.get();
-        } else if (value.getClass().isArray()) {
-            final int[] result = new int[Array.getLength(value)];
-            for (int i = 0; i < result.length; i++) {
-                result[i] = intConverter.applyAsInt(String.valueOf(Array.get(value, i)));
-            }
-            return result;
-        } else if (value.getClass().isAssignableFrom(Iterable.class)) {
-            return StreamSupport.stream(((Iterable<?>)value).spliterator(), false)
-                    .mapToInt(intConverter).toArray();
-        } else { // as a last resort with transform it into an String and try to parse an array out of it.
-            return Stream.of(String.valueOf(value).trim().replaceAll("\\[|\\]", "")
-                    .split(VCFConstants.INFO_FIELD_ARRAY_SEPARATOR))
-                    .mapToInt(intConverter).toArray();
-        }
-    }
-
-
-    /**
-     * Returns an attribute as a string.
-     * @param genotype the source genotype.
-     * @param key the attribute key.
-     * @param defaultValue default value to return in case the attribute does not have a value defined.
-     * @return the string version of the attribute value.
-     * @throws IllegalArgumentException if {@code genotype} or {@code key} is {@code null}.
-     */
-    public static String getAttributeAsString(final Genotype genotype, final String key, final String defaultValue) {
-        Utils.nonNull(genotype);
-        Utils.nonNull(key);
-        final Object value = genotype.getExtendedAttribute(key);
-        if (value == null) {
-            return defaultValue;
-        } else {
-            return String.valueOf(value);
-        }
-    }
-
-    /**
-     * Returns an attribute as an integer.
-     * @param genotype the source genotype.
-     * @param key the attribute key.
-     * @param defaultValue default value to return in case the attribute does not have a value defined.
-     * @return the string version of the attribute value.
-     * @throws IllegalArgumentException if {@code genotype} or {@code key} is {@code null}.
-     * @throws NumberFormatException if the annotation exists, and its string transformation cannot be convert into an integer.
-     */
-    public static int getAttributeAsInt(final Genotype genotype, final String key, final int defaultValue) {
-        Utils.nonNull(genotype);
-        Utils.nonNull(key);
-        final Object value = genotype.getExtendedAttribute(key);
-        if (value == null) {
-            return defaultValue;
-        } else {
-            try {
-                return Integer.parseInt(String.valueOf(value));
-            } catch (final NumberFormatException ex) {
-                throw new NumberFormatException(String.format("attribute '%s' does not have a valid integer value: '%s'", key, String.valueOf(value)));
-            }
-        }
-    }
-
-    /**
-     * Returns an attribute as a double.
-     * @param genotype the source genotype.
-     * @param key the attribute key.
-     * @param defaultValue default value to return in case the attribute does not have a value defined.
-     * @return the string version of the attribute value.
-     * @throws IllegalArgumentException if {@code genotype} or {@code key} is {@code null}.
-     * @throws NumberFormatException if the annotation exists, and its string transformation cannot be convert into a double.
-     */
-    public static double getAttributeAsDouble(final Genotype genotype, final String key, final double defaultValue) {
-        Utils.nonNull(genotype);
-        Utils.nonNull(key);
-        final Object value = genotype.getExtendedAttribute(key);
-        if (value == null) {
-            return defaultValue;
-        } else {
-            try {
-                return Double.parseDouble(String.valueOf(value));
-            } catch (final NumberFormatException ex) {
-                throw new NumberFormatException(String.format("attribute '%s' does not have a valid double value: '%s'", key, String.valueOf(value)));
-            }
-        }
-    }
-
-    //copied from htsjdk.variant.variantcontext.CommonInfo.getAttributeAsList for simplicity
-    //maybe we should expose this as a static method in htsjdk?
-    @SuppressWarnings("unchecked")
-    public static List<Object> attributeToList(final Object attribute){
-        if ( attribute == null ) return Collections.emptyList();
-        if ( attribute instanceof List) return (List<Object>)attribute;
-        if ( attribute.getClass().isArray() ) {
-            if (attribute instanceof int[]) {
-                return Arrays.stream((int[])attribute).boxed().collect(Collectors.toList());
-            } else if (attribute instanceof double[]) {
-                return Arrays.stream((double[])attribute).boxed().collect(Collectors.toList());
-            }
-            return Arrays.asList((Object[])attribute);
-        }
-        if (attribute instanceof String) {
-            return new ArrayList<>(Arrays.asList((Object[])((String)attribute).split(",")));
-        }
-        return Collections.singletonList(attribute);
-    }
-
-    /**
-     * Instead of using the GQ value, it re-calculates the quality from the PL so that it does not need
-     * to be bounded by an artificial maximum such as the standard GQ = 99.
-     * @param genotype the PL source genotype.
-     *
-     * @return never {@code null}, but {@link Double#NaN} if there is no PLs.
-     * @throws IllegalArgumentException if {@code genotype} is {@code null}
-     */
-    public static double calculateGenotypeQualityFromPLs(final Genotype genotype) {
-        Utils.nonNull(genotype);
-        if (!genotype.hasPL()) {
-            return Double.NaN;
-        } else {
-            final int GQ = MathUtils.secondSmallestMinusSmallest(genotype.getPL(), -1);
-            return GQ < 0 ? Double.NaN : (double) GQ;
-        }
-    }
-
-    /**
-     * Instead of using the GQ value, it re-calculates the quality from the PL so that it does not need
-     * to be bounded by an artificial maximum such as the standard GQ = 99.
-     * @param builder where to set the genotypes.
-     * @param genotype the PL source genotype.
-     *
-     * @throws IllegalArgumentException if {@code genotype} is {@code null}
-     */
-    public static void setGenotypeQualityFromPLs(final GenotypeBuilder builder, final Genotype genotype) {
-        final double gq = calculateGenotypeQualityFromPLs(genotype);
-        if (Double.isNaN(gq)) {
-            builder.noGQ();
-        } else {
-            builder.GQ((int) Math.floor(gq));
-        }
-    }
-
-    /**
-     * Gets an attribute value by transforming the string its string representation using an translation function.
-     * @param g the source genotype.
-     * @param key the attribute key.
-     * @param translate the function to translate from an string to the return class of interest.
-     * @param defaultValue the default value to return in case the attribute is not defined or is declared as missing.
-     * @param <T> the return object type.
-     * @return {@code null} or an instance of the class {@link T}
-     */
-    public static <T> T getAttributeAsObject(final Genotype g, final String key, final Function<String, T> translate,
-                                                                              final T defaultValue) {
-        Utils.nonNull(g);
-        Utils.nonNull(key);
-        final Object value = g.getExtendedAttribute(key);
-        if (value == null || VCFConstants.MISSING_VALUE_v4.equals(value)) {
-            return defaultValue;
-        } else {
-            return translate.apply(String.valueOf(value));
-        }
-    }
 
     /**
      * Get the pileup of reads covering a locus.  This is useful, for example, in VariantWalkers, which work on
